+++ src/components/CreateGoal/CreateGoal.js
@@ -8,7 +8,7 @@ import "./styles.css";
 import { useSelector, useDispatch } from "react-redux";
 
 
-//import Expenses from '../CreateGoal/Expenses';
+import Expenses from '../CreateGoal/Expenses';
 
 
 
@@ -23,79 +23,7 @@ export default function CreateGoal() {
 
     const [buttonClick, setButtonClick] = useState(true);
   
-    
     const dispatch = useDispatch();
-
-
-    // const handleChange = (value) => {
-    //     setState({ value });
-    // };
-
-    // const postGoals = (event) => {
-    //     event.preventDefault(event);
-
-    //     dispatch({
-    //         type: "POST_GOALS",
-    //         payload: 
-    //         {
-    //             name: goal,
-    //             reasons: reasons,
-    //         },
-    //     });
-    // };
-
-
-    let data = [{}];
-
-    const [goalData, setGoalData] = useState(data);
-    const [addFormData, setAddFormData] = useState({
-        expense: '',
-        price: '',
-        notes: ''
-    })
-
-    const handleAddFormChange = (event) => {
-        event.preventDefault(event);
-
-        // Get the name attribute from TextField on handleAddSubmit form (expense, price, notes)
-        const fieldName = event.target.getAttribute('name');
-
-        // Get the value the user enters into the TextField (expense, price, notes)
-        const fieldValue = event.target.value;
-
-        // Spread operator to copy existing form data
-        const newFormData = { ...addFormData };
-
-        // Update the object with the new value the user inputs
-        newFormData[fieldName] = fieldValue;
-
-        // Set newFormData to state
-        setAddFormData(newFormData)
-    }
-
-    const handleBudgetSubmit = (event) => {
-        event.preventDefault(event);
-
-        // Create new object from addFormData
-        const newGoalData = {
-            expense: addFormData.expense,
-            price: addFormData.price,
-            notes: addFormData.notes
-        };
-
-        // Create new array to avoid mutating the state 
-        const newGoalsData = [...goalData, newGoalData];
-        setGoalData(newGoalsData);
-
-        dispatch({
-            type: "POST_NEW_EXPENSE",
-            payload: newGoalData
-            // {
-            //     name: goal,
-            //     reasons: reasons,
-            // },
-        });
-    }
 
     const addExpenses = () => {
         setButtonClick(!buttonClick);
@@ -117,43 +45,7 @@ export default function CreateGoal() {
 
             return (
                 <div>
-                    <table class="expenseTable">
-                        <thead>
-                            <tr>
-                                <th>Expense</th>
-                                <th>Price</th>
-                                <th>Notes</th>
-                            </tr>
-                        </thead>
-                        <tbody>
-                            {goalData.map((data) => (
-                                <tr>
-                                    <td className="expenseTableData">{data.expense}</td>
-                                    <td className="expenseTableData">{data.price}</td>
-                                    <td className="expenseTableData">{data.notes}</td>
-                                </tr>
-                            ))}
-                        </tbody>
-
-                    </table>
-                    <div>
-                        <TextField
-                            placeholder="Expense"
-                            name="expense"
-                            onChange={handleAddFormChange}
-                        />
-                        <TextField
-                            placeholder="Price"
-                            name="price"
-                            onChange={handleAddFormChange}
-                        />
-                        <TextField
-                            placeholder="Notes"
-                            name="notes"
-                            onChange={handleAddFormChange}
-                        />
-                        <Button onClick={handleBudgetSubmit} type="submit" variant="contained">Add Expense</Button>
-                    </div>
+                    <Expenses />
                 </div>
             )
         }
@@ -183,13 +75,6 @@ export default function CreateGoal() {
                     value={reasons.value}
                     onChange={(event) => setReasons(event.target.value)}
                 />
-                {/* <Button
-                    type="submit"
-                    size="small"
-                    variant="contained"
-                >
-                    Add Expenses
-                </Button> */}
                 <Button
                     variant="contained"
                     onClick={addExpenses}
@@ -210,13 +95,6 @@ export default function CreateGoal() {
                     modules={modules}
                     formats={formats}
                 /> */}
-
-                
-                {/* {goalStuffIsCool <= 0 ?
-                    <div>Completed a goal first :)s</div>
-                    :       
-                  <Expenses  />
-                } */}
             </form>
         </div>
     );
